+++ includes/os/class.WINNT.inc.php
@@ -187,11 +187,17 @@ class WINNT extends OS
         }
         if ($buffer && isset($buffer[0])) {
             if (isset($buffer[0]['CodeSet'])) {
-                if ($buffer[0]['CodeSet'] == 949) {
-                    $this->_codepage = 'EUC-KR';
+                $codeset = $buffer[0]['CodeSet'];
+                if ($codeset == 932) {
+                    $codename = ' (SJIS)';
+                } elseif ($codeset == 949) {
+                    $codename = ' (EUC-KR)';
+                } elseif ($codeset == 950) {
+                    $codename = ' (BIG-5)';
                 } else {
-                    $this->_codepage = 'windows-'.$buffer[0]['CodeSet'];
+                    $codename = '';
                 }
+                $this->_codepage = 'windows-'.$codeset.$codename;
             }
             if (isset($buffer[0]['OSLanguage'])) {
                 $lang = "";